+++ scripts/shared/upload-examples-to-s3.js
