+++ scripts/shared/upload-sdk-to-s3.js
